requirements.txt
beautifulsoup4
feedparser
flask
openai
psycopg2-binary
python-dotenv
requests
